requirements.txt
cachecontrol==0.11.6
canonicalwebteam.versioned-static==1.0.1
canonicalwebteam.get-feeds==0.1.7
lockfile==0.12.2
Django==1.11.1
whitenoise==3.3.0
django-template-finder-view==0.2
django-json-redirects==0.3
django-static-root-finder==0.1
django-asset-server-url==0.1
feedparser==5.2.1
requests==2.10.0
requests-cache==0.4.12
python-dateutil==2.7.0
